cim-poc/src/python/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asn1crypto==1.4.0
async-timeout==4.0.2
attrs==21.4.0
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
clients==1.3
config==0.5.1
cryptography==36.0.1
cymem==2.0.6
et-xmlfile==1.1.0
frozenlist==1.3.0
future==0.18.2
fuzzywuzzy==0.18.0
h11==0.12.0
httpcore==0.14.5
httpx==0.22.0
idna==3.3
Jinja2==3.0.3
joblib==1.1.0
jsonpickle==2.1.0
langcodes==3.3.0
MarkupSafe==2.0.1
multidict==6.0.2
murmurhash==1.0.6
nltk==3.7
numpy==1.22.3
openpyxl==3.0.9
oscrypto==1.2.1
packaging==21.3
pandas==1.4.1
pandarallel==1.6.1
pathy==0.6.1
preshed==3.0.6
probableparsing==0.0.1
pyarrow==6.0.1
pycparser==2.21
pycryptodomex==3.14.0
pydantic==1.8.2
pygeocodio>=1.0.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
python-crfsuite==0.9.7
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
rfc3986==1.5.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
snowflake-connector-python==2.7.8
SQLAlchemy==1.4.31
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.1.0
tqdm==4.62.3
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.8
usaddress==0.5.10
utility==1.0
wasabi==0.9.0
yarl==1.7.2
hyperopt==0.2.7
toolz==0.11.2
pyyaml==6.0
python-json-logger==2.0.2
timeout-decorator==0.5.0
sparse_dot_mkl==0.8.1
pytest==7.1.1
uszipcode==1.0.1

python-dotenv~=0.20.0
lightgbm~=3.3.2